+++ miniGame/Otp_table.h
@@ -10,11 +10,11 @@ public:
 
 	friend std::ostream& operator<<(std::ostream& os, const Otp_table& T);
 
-	Otp_table(size_t n);
+	Otp_table(int n);
 
 private:
 	std::vector<std::vector<std::string>> table;
 	std::vector<size_t>len;
-	size_t tot_len = 0;
+	int tot_len = 0;
 };
